useRecoilSnapshotAndSubscribe() hook (#311)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/311

Add `useRecoilSnapshotAndSubscribe()` hook.  This was originally part of [this proposal](https://fb.quip.com/UwKaAGjK0L0E).  We removed it thinking that `useRecoilTransactionObserver()` would be sufficient.  However, we need this hook to be able to synchronously get a snapshot during render in order to support server-side rendering.  We could drop `useRecoilTransactionObserver()` and only use this hook.  However, that one has a more convenient API for persistence/debug tools.  So, keep both as a convenience at the expense of a larger API surface area.

```
declare function useRecoilSnapshotAndSubscribe(): Snapshot;
```

Label hook as `useRecoilSnapshotAndSubscribe()` to make more obvious to the user that using this hook will also subscribe that component to all transactions.  It should be used sparingly.

In the future, we plan to add a parameter to `useRecoilSnapshotAndSubscribe()` and `useRecoilTransactionObserver()` to control debouncing or condition when to update the subscription.

Reviewed By: csantos42

Differential Revision: D21783137

fbshipit-source-id: 51096db1fcd50b59017a9b2ef440f1d9418cbbb9