<Link> component (#379)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/379

Provide new Link components for Recoil URL persistence based on [this proposal](https://fb.quip.com/UwKaAGjK0L0E).  There are two variants:

# `<LinkToRecoilSnapshot>`
A Link component based on the provided `uriFromSnapshot` mapping of a URI from a Recoil Snapshot.

The Link element renders an anchor element.  But instead of an href, use a `snapshot` property.  When clicked, the Link element updates the current state to the snapshot without loading a new document.

The href property of the anchor will set using `uriFromSnapshot`.  This allows users to copy the link, choose to open in a new tab, &c.

If an `onClick` handler is provided, it is called before the state transition and may call preventDefault on the event to stop the state transition.

# `<LinkToRecoilStateChange>`
A Link component based on the provided `uriFromSnapshot` mapping of a URI from a Recoil Snapshot.

The Link element renders an anchor element.  But instead of an href, use a `stateChange` property.  When clicked, the Link element updates the current state based on the `stateChange` callback without loading a new document. `stateChange` is a function which takes a `MutableSnapshot` that can be used to read the current state and set or update any changes.

The href property of the anchor will set using `uriFromSnapshot`.  This allows users to copy the link, choose to open in a new tab, &c.

If an `onClick` handler is provided, it is called before the state transition and may call preventDefault on the event to stop the state transition.

*Note that, because the link renders the href based on the current state snapshot it is re-rendered whenever any state change is made.  Keep the performance implications of this in mind.*

Reviewed By: habond

Differential Revision: D21784015

fbshipit-source-id: 9cd6ff2a9fa92bb094175472416a639a2a4f8142