Fix TS lint error and improve typing of DefaultValue (#492)
Summary:
`DefaultValue` was defined as an empty class, which means it was behaving as `{}`, so it was allowing almost any type to be assigned to it. This PR introduces a private tag to solve the issue.

Also, this PR introduces a minor change to `SnapshotID` to conform with tslint

fixes https://github.com/facebookexperimental/Recoil/issues/487

Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/492

Reviewed By: drarmstr

Differential Revision: D22673551

Pulled By: csantos42

fbshipit-source-id: c19b323379bef19296bbee5642889ad8875246c2