More robust fix for DOM updates during flushes for tests. (#576)
Summary:
If there are updates during a flush which affect the React DOM then they need to be wrapped in an `act()` call during Jest tests.  This often only impacts testing with the OSS package and not WWW.  Instead of discovering later that this is an issue just always call `act()` from within all `flushPromisesAndTimers()` calls.

Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/576

Test Plan: test in both www and OSS with jest unit tests

Reviewed By: davidmccabe

Differential Revision: D23707457

Pulled By: drarmstr

fbshipit-source-id: 1f13968d1b8eab267f1d2b16e409ac49e5d0b0a1