Use immutable map structures
Summary:
Use hashed-array-mapped tries instead of built-in Map for atomValues, to avoid copying the map. This should dramatically increase speed for large numbers of atoms.

In a benchmark I performed, setting a value in a HAMT (just the data structure, not all of Recoil) compared with copying a map and setting the value is:

* 3.7× faster for 10 entries
* 28× faster for 100 entries
* 325× faster for 1,000 entries
* 3,026× faster for 10,000 entries

Reading values is slower:

* 5× slower for 10 entries
* 5× slower for 100 entries
* 6× slower for 1,000 entries
* 8× slower for 10,000 entries

However, in Recoil, we generally only read once per write (since the read value is cached elsewhere until the next write), so this should still be a win overall.

Reviewed By: drarmstr

Differential Revision: D25487986

fbshipit-source-id: cea056cc298520f1b54f9fb12c896086c4ef82c6