Add error message for useRecoilCallback() misuse (#870)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/870

Add a slightly better error message when `useRecoilCallback()` is misused when a user calls it with a function that just returns a value instead a function that takes the callback interface and returns a function that matches the signature of the callback we are producing.

Before: `Uncaught TypeError: fn(...) is not a function`
After: `Uncaught Error: useRecoilCallback() accepts a function parameter which returns a function that matches the signature of the callback function.`

Flow and TypeScript will catch this, but this runtime check may help those not using a type system. (#851)

Reviewed By: davidmccabe

Differential Revision: D26219475

fbshipit-source-id: 7d19cbad36fe4f46b2d14e4fdd3efcfce20eb450