RecoilLoadable.all() accepts literal values and Promises (#1442)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1442

Allow the `RecoilLoadable.all()` factory to accept literal values and async Promises in addition to Loadables.  This makes the API more convenient to work with, especially when working with nullable loadables.

This is similar to `Promise.all()` which accepts an array of either Promises or literal values.

## Example
```
RecoilLoadable.all({
  literal: 'A',
  loadable: RecoilLoadable.of('B'),
  promise: Promise.resolves('C'),
});
```
resolves to: `{literal: 'A', loadable: 'B', promise: 'C'}`

Update TypeScript.

Reviewed By: habond

Differential Revision: D32526729

fbshipit-source-id: 46dada2e39e4f03178b61357ac804377b006a3f5