Optimize snapshot cloning overhead for callbacks (#1533)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/1533

If a user has a recoil callback, such as from `useRecoilCallback()`, and uses it to get a snapshot many times in a loop it could incur overhead for cloning the snapshot.  Optimize this by caching the cloned snapshot and only cloning a new one if the store or state has changed.

Reviewed By: habond

Differential Revision: D33490900

fbshipit-source-id: 093b2aa498337bb2abc8eebe54a72e1a130b3469