Avoid unhandled promise rejection in callbacks (#2075)
Summary:
The creation of a finally branch of the promise in `useRecoilCallback` causes unhandled promise rejections, despite the fact that the promise is actually handled downstream. This change returns the finally branch instead of the original promise, so there's no divergence in the handling.

For example, the following selector currently "throws" twice, once caught/logged and once not caught:

```javascript
const testState = selector({
  key: 'testState',
  get: ({ getCallback }) =>
    getCallback(() => async () => {
      throw 'test';
    }),
});

function Component() {
  const test = useRecoilValue(testState);

  useEffect(() => {
    test().then(
      () => console.log('OK'),
      (error) => console.log(error),
    );
  }, [test]);

  // ...
}
```

Pull Request resolved: https://github.com/facebookexperimental/Recoil/pull/2075

Reviewed By: drarmstr

Differential Revision: D40573295

Pulled By: wd-fb

fbshipit-source-id: 2683fbb2d881a61ab800c114d0c72997aa25dd3a